게시글 수정

+++ src/main/java/com/example/project/springbootproject/repository/BoardRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BoardRepository extends JpaRepository<Board, Integer> {
+public interface BoardRepository extends JpaRepository<Board, Long> {
     List<Board> findAllByUsername(String username);
 
-    Board findById(int boardId);
+    Board findById(long boardId);
 
 }
